+++ Source/OpenStreetMapModule/Private/OpenStreetNode.cpp
@@ -8,11 +8,11 @@ FOpenStreetNode::FOpenStreetNode() : Id(0), Location()
 {
 }
 
-FOpenStreetNode::FOpenStreetNode(int64 Id) : Id(Id), Location()
+FOpenStreetNode::FOpenStreetNode(int32 Id) : Id(Id), Location()
 {
 }
 
-FOpenStreetNode::FOpenStreetNode(int64 Id, FLatLng Location) : Id(Id), Location(Location)
+FOpenStreetNode::FOpenStreetNode(int32 Id, FLatLng Location) : Id(Id), Location(Location)
 {
 }
